# ingestion-server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/auth/register":{"post":{"operationId":"register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}}},"responses":{"200":{"description":"User Registered Successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponseDTO"}}}},"400":{"description":"Bad Request"}},"summary":"Register user to get APIKEY and APISECRET","tags":["Auth"]}},"/api/auth/token":{"post":{"operationId":"generateToken","parameters":[],"responses":{"200":{"description":"Token Generated Successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateTokenResponseDTO"}}}},"400":{"description":"Bad Request"}},"summary":"Generate JWT token with APIKEY and APISECRET","tags":["Auth"]}},"/api/auth/keys/regenerate":{"post":{"operationId":"regenerateKeys","parameters":[],"responses":{"200":{"description":"Keys Regenerated Successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponseDTO"}}}},"400":{"description":"Bad Request"}},"summary":"Regenerate APIKEY and APISECRET","tags":["Auth"]}},"/api/events":{"post":{"description":"Singular endpoint to create either a single object event entry or an array of object event entries","operationId":"createInterationEvent","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInteractionEventBatchDTO"}}}},"responses":{"200":{"description":"Event Created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/"}}}},"400":{"description":"Bad Request"}},"summary":"Create new single or batch interation events","tags":["Interaction Event"]}},"/api/sync/delta":{"get":{"description":"Fetches the modified data from the timestamp provide for a particular project","operationId":"deltaSync","parameters":[{"name":"modifiedSince","required":true,"in":"query","description":"Time of the last sync","schema":{"format":"date-time","example":"2026-05-10T00:00:00.000Z","type":"string"}},{"name":"projectId","required":true,"in":"query","description":"Unique UUID of the project the data belongs to","schema":{"example":"3fa85f64-5717-4562-b3fc-2c963f66afa6","type":"string"}},{"name":"limit","required":false,"in":"query","description":"Max number of event items returned","schema":{"example":5000,"type":"number"}}],"responses":{"200":{"description":"Modified Events Fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/"}}}}},"summary":"Fetch latest modified data","tags":["Delta Sync"]}},"/api/recommendations/bulk":{"post":{"description":"Records the recommended data for a user of a particular project","operationId":"recordRecommendations","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationBulkDTO"}}}},"responses":{"200":{"description":"Recommendations Recorded successfully"}},"summary":"Record recommended data","tags":["Recommendations"]}},"/api/recommendations/{userId}":{"get":{"description":"Fetches the recommended data for a user of a particular project","operationId":"getRecommendation","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Recommendations Fetched successfully"}},"summary":"Fetches recommended data","tags":["Recommendations"]}}},"info":{"title":"OpenRec Ingestion-Server","description":"","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:4002","description":"Local Ingestion Server"}],"components":{"schemas":{"RegisterDTO":{"type":"object","properties":{"username":{"type":"string","description":"A unique username","example":"exarl"},"email":{"type":"string","description":"Email of the user","example":"example@gmail.com"},"password":{"type":"string","description":"Password of the user","example":"12345"},"confirmPassword":{"type":"string","description":"The same password as the previous","example":"12345"}},"required":["username","email","password","confirmPassword"]},"RegisterResponseDTO":{"type":"object","properties":{"apiKey":{"type":"string","description":"Unique key given after registration","example":"ingest_apk_xxxxxxxxxxxxxx"},"apiSecret":{"type":"string","description":"Unique secret given after registration","example":"ingest_asec_xxxxxxxxxxxxx"}},"required":["apiKey","apiSecret"]},"GenerateTokenResponseDTO":{"type":"object","properties":{"token":{"type":"string","description":"Jwt token","example":"efheytjf.xxxxxxxxxxxxxx.skfjdksf"}},"required":["token"]},"CreateInteractionEventDTO":{"type":"object","properties":{"projectId":{"type":"string","description":"Unique UUID of the project","example":"3fa85f64-5717-4562-b3fc-2c963f66afa6"},"userId":{"type":"string","description":"Unique ID of the user","example":"user-1"},"itemId":{"type":"string","description":"Unique ID of the item","example":"234489"},"eventType":{"type":"string","enum":["VIEW","CLICK","LIKE","ADD_TO_CART","PURCHASE","RATING","DISLIKE"],"description":"Type of event carried out by the user on the item"},"interactionValue":{"type":"number","description":"The alloted value/score assigned to the item","example":5}},"required":["projectId","userId","itemId","eventType","interactionValue"]},"CreateInteractionEventBatchDTO":{"type":"object","properties":{"events":{"description":"The batch list of Events","type":"array","items":{"$ref":"#/components/schemas/CreateInteractionEventDTO"}}},"required":["events"]},"RecommendationBulkDTO":{"type":"object","properties":{"recommendations":{"type":"array","description":"All Available Recommendations","example":[{"projectId":"3fa85f64-5717-4562-b3fc-2c963f66afa6","userId":"user-1","recommendedItemIds":[{"itemId":"234489","score":0.95}]},{"projectId":"3fa85f64-5717-4562-b3fc-2c963f66afa6","userId":"user-2","recommendedItemIds":[{"itemId":"234","score":0.75},{"itemId":"338","score":0.9}]}]}},"required":["recommendations"]}}}}